Use yew::Callbacks instead of custom (for now)
Ultimate goal for #74 is to move away from yew dependency, so originally
used custom callbacks instead of yew's.  But in working on 
websocket/webtransport connection (in another branch) it's currently
deeply tied to yew Callback.

So for now, for consistency it makes sense to
use yew Callbacks everywhere.

Later on in working on #74 will try to come up with a consistent approach
to extracting the yew Callback dependency.